auctions: make uint48 arithmetic actually safe

+++ src/flap.sol
@@ -73,7 +73,7 @@ contract Flapper is DSNote {
     }
 
     // --- Math ---
-    function add(uint x, uint y) internal pure returns (uint z) {
+    function add(uint48 x, uint48 y) internal pure returns (uint48 z) {
         z = x + y;
         require(z >= x);
     }
@@ -93,7 +93,7 @@ contract Flapper is DSNote {
         bids[id].bid = bid;
         bids[id].lot = lot;
         bids[id].guy = msg.sender; // configurable??
-        bids[id].end = uint48(add(now, tau));
+        bids[id].end = add(uint48(now), tau);
         bids[id].gal = gal;
 
         dai.move(msg.sender, this, lot);
@@ -114,7 +114,7 @@ contract Flapper is DSNote {
 
         bids[id].guy = msg.sender;
         bids[id].bid = bid;
-        bids[id].tic = uint48(add(now, ttl));
+        bids[id].tic = add(uint48(now), ttl);
     }
     function deal(uint id) public note {
         require(bids[id].tic < now && bids[id].tic != 0 ||
